+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+Version 0.2.0
+===============
+- Added support for querying data from [Trials Report](https://trials.report) ([Github](https://github.com/DestinyTrialsReport/DestinyTrialsReport))
+  - Top used weapons and categories
+  - Personal performance stats for current week
+  - Personal top used weapons
+  - Current map
+
 Version 0.1.0
 ===============
 - Initial "minimum viable product" release.